refactor(background-worker): decompose BackgroundActionOrchestratorService into partials
Extract handler invocation logic and type definitions into dedicated partials. Reduces orchestrator main class to core orchestration responsibilities.

Ultraworked with [Sisyphus](https://github.com/code-yeongyu/oh-my-openagent)
